+++ extensions/qqbot/src/bridge/narrowing.ts
@@ -1,0 +1,28 @@
+import type { OpenClawConfig } from "openclaw/plugin-sdk/config-runtime";
+import type { PluginRuntime } from "openclaw/plugin-sdk/core";
+import type { GatewayAccount } from "../engine/types.js";
+import type { ResolvedQQBotAccount } from "../types.js";
+
+/**
+ * Map resolved plugin account to the engine gateway account shape (single assertion on nested config).
+ */
+export function toGatewayAccount(account: ResolvedQQBotAccount): GatewayAccount {
+  return {
+    accountId: account.accountId,
+    appId: account.appId,
+    clientSecret: account.clientSecret,
+    markdownSupport: account.markdownSupport,
+    systemPrompt: account.systemPrompt,
+    config: account.config as GatewayAccount["config"],
+  };
+}
+
+/**
+ * Persist OpenClaw config through the injected plugin runtime (typed entry point).
+ */
+export async function writeOpenClawConfigThroughRuntime(
+  runtime: PluginRuntime,
+  cfg: OpenClawConfig,
+): Promise<void> {
+  await runtime.config.writeConfigFile(cfg);
+}
