+++ extensions/qqbot/src/channel.ts
@@ -10,6 +10,8 @@ import {
   DEFAULT_ACCOUNT_ID,
   resolveQQBotAccount,
 } from "./bridge/config.js";
+import type { GatewayContext } from "./bridge/gateway.js";
+import { toGatewayAccount, writeOpenClawConfigThroughRuntime } from "./bridge/narrowing.js";
 import { getQQBotRuntime } from "./bridge/runtime.js";
 import { qqbotSetupWizard } from "./bridge/setup/surface.js";
 import { qqbotChannelConfigSchema } from "./config-schema.js";
@@ -33,6 +35,12 @@ function loadGatewayModule(): Promise<typeof import("./bridge/gateway.js")> {
 
 const EXEC_APPROVAL_COMMAND_RE =
   /\/approve(?:@[^\s]+)?\s+[A-Za-z0-9][A-Za-z0-9._:-]*\s+(?:allow-once|allow-always|always|deny)\b/i;
+
+function persistAccountCredentialSnapshot(account: ResolvedQQBotAccount): void {
+  if (account.appId && account.clientSecret) {
+    saveCredentialBackup(account.accountId, account.appId, account.clientSecret);
+  }
+}
 
 function shouldSuppressLocalQQBotApprovalPrompt(params: {
   cfg: OpenClawConfig;
@@ -119,7 +127,13 @@ export const qqbotPlugin: ChannelPlugin<ResolvedQQBotAccount> = {
       await loadGatewayModule();
       const account = resolveQQBotAccount(cfg, accountId);
       const { sendText } = await import("./engine/messaging/outbound.js");
-      const result = await sendText({ to, text, accountId, replyToId, account: account as never });
+      const result = await sendText({
+        to,
+        text,
+        accountId,
+        replyToId,
+        account: toGatewayAccount(account),
+      });
       return {
         channel: "qqbot" as const,
         messageId: result.messageId ?? "",
@@ -137,7 +151,7 @@ export const qqbotPlugin: ChannelPlugin<ResolvedQQBotAccount> = {
         mediaUrl: mediaUrl ?? "",
         accountId,
         replyToId,
-        account: account as never,
+        account: toGatewayAccount(account),
       });
       return {
         channel: "qqbot" as const,
@@ -163,11 +177,7 @@ export const qqbotPlugin: ChannelPlugin<ResolvedQQBotAccount> = {
               appId: backup.appId,
               clientSecret: backup.clientSecret,
             });
-            const runtime = getQQBotRuntime();
-            await runtime.config.replaceConfigFile({
-              nextConfig: nextCfg,
-              afterWrite: { mode: "auto" },
-            });
+            await writeOpenClawConfigThroughRuntime(getQQBotRuntime(), nextCfg);
             cfg = nextCfg;
             account = resolveQQBotAccount(nextCfg, account.accountId);
             log?.info(
@@ -195,7 +205,7 @@ export const qqbotPlugin: ChannelPlugin<ResolvedQQBotAccount> = {
         abortSignal,
         cfg,
         log,
-        channelRuntime: ctx.channelRuntime as never,
+        channelRuntime: ctx.channelRuntime as GatewayContext["channelRuntime"],
         onReady: () => {
           log?.info(`[qqbot:${account.accountId}] Gateway ready`);
           ctx.setStatus({
@@ -206,9 +216,7 @@ export const qqbotPlugin: ChannelPlugin<ResolvedQQBotAccount> = {
           });
           // Snapshot credentials so we can recover from the next hot
           // upgrade that might wipe openclaw.json mid-flight.
-          if (account.appId && account.clientSecret) {
-            saveCredentialBackup(account.accountId, account.appId, account.clientSecret);
-          }
+          persistAccountCredentialSnapshot(account);
         },
         onResumed: () => {
           log?.info(`[qqbot:${account.accountId}] Gateway resumed`);
@@ -218,9 +226,7 @@ export const qqbotPlugin: ChannelPlugin<ResolvedQQBotAccount> = {
             connected: true,
             lastConnectedAt: Date.now(),
           });
-          if (account.appId && account.clientSecret) {
-            saveCredentialBackup(account.accountId, account.appId, account.clientSecret);
-          }
+          persistAccountCredentialSnapshot(account);
         },
         onError: (error) => {
           log?.error(`[qqbot:${account.accountId}] Gateway error: ${error.message}`);
@@ -238,11 +244,7 @@ export const qqbotPlugin: ChannelPlugin<ResolvedQQBotAccount> = {
       );
 
       if (changed) {
-        const runtime = getQQBotRuntime();
-        await runtime.config.replaceConfigFile({
-          nextConfig: nextCfg as OpenClawConfig,
-          afterWrite: { mode: "auto" },
-        });
+        await writeOpenClawConfigThroughRuntime(getQQBotRuntime(), nextCfg as OpenClawConfig);
       }
 
       const resolved = resolveQQBotAccount((changed ? nextCfg : cfg) as OpenClawConfig, accountId);
