+++ extensions/qqbot/src/config-schema.ts
@@ -24,6 +24,7 @@ const QQBotSttSchema = z
   .strict()
   .optional();
 
+/** When `true`, same as `mode: "partial"` and `c2cStreamApi: true` for C2C. Object form kept for legacy configs. */
 const QQBotStreamingSchema = z
   .union([
     z.boolean(),
@@ -31,7 +32,7 @@ const QQBotStreamingSchema = z
       .object({
         /** "partial" (default) enables block streaming; "off" disables it. */
         mode: z.enum(["off", "partial"]).default("partial"),
-        /** Use QQ C2C official stream_messages API (single-message updates). */
+        /** Use QQ C2C official stream_messages API; legacy, prefer `streaming: true`. */
         c2cStreamApi: z.boolean().optional(),
       })
       .passthrough(),
