+++ extensions/qqbot/src/engine/messaging/outbound.ts
@@ -1,901 +1,82 @@
-import * as fs from "node:fs";
-import * as path from "node:path";
-import { UPLOAD_PREPARE_FALLBACK_CODE } from "../api/retry.js";
-import { MediaFileType } from "../types.js";
-import type { OutboundAudioPort } from "../adapter/audio.port.js";
+/**
+ * Outbound messaging — aggregates reply limits, audio port, media sends, and text orchestration.
+ */
+
+export { setOutboundAudioPort } from "./outbound-audio-port.js";
+export type {
+  OutboundContext,
+  MediaOutboundContext,
+  OutboundResult,
+  OutboundErrorCode,
+  MediaTargetContext,
+} from "./outbound-types.js";
+export { OUTBOUND_ERROR_CODES, DEFAULT_MEDIA_SEND_ERROR } from "./outbound-types.js";
+
+export {
+  checkMessageReplyLimit,
+  recordMessageReply,
+  getMessageReplyStats,
+  getMessageReplyConfig,
+  MESSAGE_REPLY_LIMIT,
+} from "./outbound-reply.js";
+export type { ReplyLimitResult } from "./outbound-reply.js";
+
+export { resolveUserFacingMediaError } from "./outbound-result-helpers.js";
+
+export {
+  buildMediaTarget,
+  parseTarget,
+  resolveOutboundMediaPath,
+  sendDocument,
+  sendPhoto,
+  sendVideoMsg,
+  sendVoice,
+} from "./outbound-media-send.js";
+
+import type { GatewayAccount } from "../types.js";
 import { formatErrorMessage } from "../utils/format.js";
+import { debugError, debugLog, debugWarn } from "../utils/log.js";
+import { normalizeMediaTags } from "../utils/media-tags.js";
+import { decodeCronPayload } from "../utils/payload.js";
+import { normalizePath } from "../utils/platform.js";
 import {
   normalizeLowercaseStringOrEmpty,
   normalizeOptionalString,
 } from "../utils/string-normalize.js";
-
-// ---- Outbound audio adapter — injected once by gateway startup ----
-
-let _audioPort: OutboundAudioPort | null = null;
-
-/**
- * Initialize the outbound audio adapter. Called once by gateway startup
- * via `adapters.outboundAudio`.
- */
-export function setOutboundAudioPort(port: OutboundAudioPort): void {
-  _audioPort = port;
-}
-
-function getAudio(): OutboundAudioPort {
-  if (!_audioPort) {
-    throw new Error("OutboundAudioPort not initialized — call setOutboundAudioPort first");
-  }
-  return _audioPort;
-}
-
-// Re-alias for use in the file.
-function audioFileToSilkBase64(p: string, f?: string[]): Promise<string | undefined> {
-  return getAudio().audioFileToSilkBase64(p, f);
-}
-function isAudioFile(p: string, m?: string): boolean {
-  // Safe to return false when adapter is unavailable — this is a type-check
-  // function called by sendMedia's dispatch logic before any audio processing.
-  try {
-    return getAudio().isAudioFile(p, m);
-  } catch {
-    return false;
-  }
-}
-function shouldTranscodeVoice(p: string): boolean {
-  return getAudio().shouldTranscodeVoice(p);
-}
-function waitForFile(p: string, ms?: number): Promise<number> {
-  return getAudio().waitForFile(p, ms);
-}
-import type { GatewayAccount } from "../types.js";
-import {
-  checkFileSize,
-  downloadFile,
-  fileExistsAsync,
-  formatFileSize,
-  getFileTypeName,
-  getImageMimeType,
-  getMaxUploadSize,
-  readFileAsync,
-} from "../utils/file-utils.js";
-import { debugError, debugLog, debugWarn } from "../utils/log.js";
-import { normalizeMediaTags } from "../utils/media-tags.js";
-import { decodeCronPayload } from "../utils/payload.js";
-import {
-  getQQBotDataDir,
-  getQQBotMediaDir,
-  isLocalPath as isLocalFilePath,
-  normalizePath,
-  resolveQQBotPayloadLocalFilePath,
-} from "../utils/platform.js";
-import { sanitizeFileName } from "../utils/string-normalize.js";
 import {
   isImageFile as coreIsImageFile,
   isVideoFile as coreIsVideoFile,
 } from "./media-type-detect.js";
-// Bridge to core/ modules — use the canonical implementations from the core
-// package so the same logic can be shared with the standalone version.
-import { ReplyLimiter, type ReplyLimitResult } from "./reply-limiter.js";
+import { isAudioFile } from "./outbound-audio-port.js";
 import {
-  sendText as senderSendText,
-  sendMedia as senderSendMedia,
+  buildMediaTarget,
+  parseTarget,
+  resolveOutboundMediaPath,
+  sendDocument,
+  sendPhoto,
+  sendVideoMsg,
+  sendVoice,
+} from "./outbound-media-send.js";
+import {
+  checkMessageReplyLimit,
+  MESSAGE_REPLY_LIMIT,
+  recordMessageReply,
+} from "./outbound-reply.js";
+import type {
+  MediaOutboundContext,
+  MediaTargetContext,
+  OutboundContext,
+  OutboundResult,
+} from "./outbound-types.js";
+import {
   initApiConfig,
   accountToCreds,
-  UploadDailyLimitExceededError,
+  sendText as senderSendText,
   type DeliveryTarget,
 } from "./sender.js";
-import { parseTarget as coreParseTarget } from "./target-parser.js";
 
-// Module-level reply limiter instance (replaces the old Map-based tracker).
-const replyLimiter = new ReplyLimiter();
-
-// Limit passive replies per message_id within the QQ Bot reply window.
-// Delegated to core/messaging/reply-limiter.ts for cross-version sharing.
-const MESSAGE_REPLY_LIMIT = 4;
-
-/** Result of the passive-reply limit check. */
-export type { ReplyLimitResult };
-
-/** Check whether a message can still receive a passive reply. */
-export function checkMessageReplyLimit(messageId: string): ReplyLimitResult {
-  return replyLimiter.checkLimit(messageId);
-}
-
-/** Record one passive reply against a message. */
-export function recordMessageReply(messageId: string): void {
-  replyLimiter.record(messageId);
-  debugLog(
-    `[qqbot] recordMessageReply: ${messageId}, count=${replyLimiter.getStats().totalReplies}`,
-  );
-}
-
-/** Return reply-tracker stats for diagnostics. */
-export function getMessageReplyStats(): { trackedMessages: number; totalReplies: number } {
-  return replyLimiter.getStats();
-}
-
-/** Return the passive-reply configuration. */
-export function getMessageReplyConfig(): { limit: number; ttlMs: number; ttlHours: number } {
-  return replyLimiter.getConfig();
-}
-
-export interface OutboundContext {
-  to: string;
-  text: string;
-  accountId?: string | null;
-  replyToId?: string | null;
-  account: GatewayAccount;
-}
-
-export interface MediaOutboundContext extends OutboundContext {
-  mediaUrl: string;
-  mimeType?: string;
-}
-
-/**
- * Stable error codes for outbound media send results.
- *
- * Unlike the free-form `error` string (which is a diagnostic aid, not a
- * contract), these codes are the programmatic signal that upper layers
- * use to decide how to react — e.g. whether to surface the original
- * `error` to the user verbatim or to fall back to a generic message.
- */
-export const OUTBOUND_ERROR_CODES = {
-  /** File exceeds the per-type upload ceiling (see `MEDIA_FILE_TYPE_INFO`). */
-  FILE_TOO_LARGE: "file_too_large",
-  /** Hit the per-day cumulative upload quota (QQ biz code 40093002). */
-  UPLOAD_DAILY_LIMIT_EXCEEDED: "upload_daily_limit_exceeded",
-} as const;
-
-export type OutboundErrorCode = (typeof OUTBOUND_ERROR_CODES)[keyof typeof OUTBOUND_ERROR_CODES];
-
-/** Default user-facing message when the underlying `error` should not leak. */
-export const DEFAULT_MEDIA_SEND_ERROR = "发送失败，请稍后重试。";
-
-export interface OutboundResult {
-  channel: string;
-  messageId?: string;
-  timestamp?: string | number;
-  error?: string;
-  /**
-   * Stable error code that callers can key on without depending on the
-   * `error` text. Absent for success or uncategorized failures.
-   */
-  errorCode?: OutboundErrorCode;
-  /** QQ Open Platform business error code (e.g. `upload_prepare` 40093002). */
-  qqBizCode?: number;
-  refIdx?: string;
-}
-
-/**
- * Convert a media send result into a user-facing message.
- *
- * Only errors that were explicitly categorized as safe to display verbatim
- * (via `errorCode` / `qqBizCode`) are forwarded; everything else collapses
- * to {@link DEFAULT_MEDIA_SEND_ERROR} so raw server errors / stack traces
- * never surface to end users.
- */
-export function resolveUserFacingMediaError(
-  result: Pick<OutboundResult, "error" | "errorCode" | "qqBizCode">,
-): string {
-  if (!result.error) {
-    return DEFAULT_MEDIA_SEND_ERROR;
-  }
-  if (result.qqBizCode === UPLOAD_PREPARE_FALLBACK_CODE) {
-    return result.error;
-  }
-  switch (result.errorCode) {
-    case OUTBOUND_ERROR_CODES.FILE_TOO_LARGE:
-    case OUTBOUND_ERROR_CODES.UPLOAD_DAILY_LIMIT_EXCEEDED:
-      return result.error;
-    default:
-      return DEFAULT_MEDIA_SEND_ERROR;
-  }
-}
-
-/**
- * Shape a {@link UploadDailyLimitExceededError} into an {@link OutboundResult}
- * with the canonical user-facing text.
- */
-function buildDailyLimitExceededResult(err: UploadDailyLimitExceededError): OutboundResult {
-  const dir = path.dirname(err.filePath);
-  const name = path.basename(err.filePath);
-  const size = formatFileSize(err.fileSize);
-  return {
-    channel: "qqbot",
-    error: `QQBot每天发送文件有累计2G的限制，如果着急的话，可以直接来我的主机copy下载，文件目录\`${dir}/${name}\`（${size}）`,
-    errorCode: OUTBOUND_ERROR_CODES.UPLOAD_DAILY_LIMIT_EXCEEDED,
-    qqBizCode: UPLOAD_PREPARE_FALLBACK_CODE,
-  };
-}
-
-/**
- * Shape a "file exceeds per-type ceiling" result with the canonical
- * user-facing text and the {@link OUTBOUND_ERROR_CODES.FILE_TOO_LARGE}
- * error code.
- */
-function buildFileTooLargeResult(fileType: MediaFileType, fileSize: number): OutboundResult {
-  const typeName = getFileTypeName(fileType);
-  const limit = getMaxUploadSize(fileType);
-  const limitMB = Math.round(limit / (1024 * 1024));
-  return {
-    channel: "qqbot",
-    error: `${typeName}过大（${formatFileSize(fileSize)}），超过了${limitMB}M，暂时不能通过QQ直接发给你。`,
-    errorCode: OUTBOUND_ERROR_CODES.FILE_TOO_LARGE,
-  };
-}
-
-/** Parse a qqbot target into a structured delivery target. */
-function parseTarget(to: string): { type: "c2c" | "group" | "channel"; id: string } {
-  const timestamp = new Date().toISOString();
-  debugLog(`[${timestamp}] [qqbot] parseTarget: input=${to}`);
-  const parsed = coreParseTarget(to);
-  debugLog(`[${timestamp}] [qqbot] parseTarget: ${parsed.type} target, ID=${parsed.id}`);
-  return parsed;
-}
-
-// Structured media send helpers shared by gateway delivery and sendText.
-
-/** Normalized target information for media sends. */
-export interface MediaTargetContext {
-  targetType: "c2c" | "group" | "channel" | "dm";
-  targetId: string;
-  account: GatewayAccount;
-  replyToId?: string;
-  /** Optional log prefix for diagnostics (e.g. streaming media sends). */
-  logPrefix?: string;
-}
-
-/** Build a media target from a normal outbound context. */
-function buildMediaTarget(ctx: {
-  to: string;
-  account: GatewayAccount;
-  replyToId?: string | null;
-}): MediaTargetContext {
-  const target = parseTarget(ctx.to);
-  return {
-    targetType: target.type,
-    targetId: target.id,
-    account: ctx.account,
-    replyToId: ctx.replyToId ?? undefined,
-  };
-}
-
-/** Return true when public URLs should be passed through directly. */
-function shouldDirectUploadUrl(account: GatewayAccount): boolean {
-  return account.config?.urlDirectUpload !== false;
-}
-
-type QQBotMediaKind = "image" | "voice" | "video" | "file" | "media";
-
-const qqBotMediaKindLabel: Record<QQBotMediaKind, string> = {
-  image: "Image",
-  voice: "Voice",
-  video: "Video",
-  file: "File",
-  media: "Media",
-};
-
-type ResolvedOutboundMediaPath = { ok: true; mediaPath: string } | { ok: false; error: string };
-type ResolveOutboundMediaPathOptions = {
-  allowMissingLocalPath?: boolean;
-  extraLocalRoots?: string[];
-};
-type SendDocumentOptions = {
-  allowQQBotDataDownloads?: boolean;
-};
-
-function isHttpOrDataSource(pathValue: string): boolean {
-  return (
-    pathValue.startsWith("http://") ||
-    pathValue.startsWith("https://") ||
-    pathValue.startsWith("data:")
-  );
-}
-
-function isPathWithinRoot(candidate: string, root: string): boolean {
-  const relative = path.relative(root, candidate);
-  return relative === "" || (!relative.startsWith("..") && !path.isAbsolute(relative));
-}
-
-function resolveMissingPathWithinMediaRoot(normalizedPath: string): string | null {
-  const resolvedCandidate = path.resolve(normalizedPath);
-  if (fs.existsSync(resolvedCandidate)) {
-    return null;
-  }
-
-  const allowedRoot = path.resolve(getQQBotMediaDir());
-  let canonicalAllowedRoot: string;
-  try {
-    canonicalAllowedRoot = fs.realpathSync(allowedRoot);
-  } catch {
-    return null;
-  }
-
-  const missingSegments: string[] = [];
-  let cursor = resolvedCandidate;
-  while (!fs.existsSync(cursor)) {
-    const parent = path.dirname(cursor);
-    if (parent === cursor) {
-      break;
-    }
-    missingSegments.unshift(path.basename(cursor));
-    cursor = parent;
-  }
-
-  if (!fs.existsSync(cursor)) {
-    return null;
-  }
-
-  let canonicalCursor: string;
-  try {
-    canonicalCursor = fs.realpathSync(cursor);
-  } catch {
-    return null;
-  }
-  const canonicalCandidate =
-    missingSegments.length > 0 ? path.join(canonicalCursor, ...missingSegments) : canonicalCursor;
-
-  return isPathWithinRoot(canonicalCandidate, canonicalAllowedRoot) ? canonicalCandidate : null;
-}
-
-function resolveExistingPathWithinRoots(
-  normalizedPath: string,
-  allowedRoots: readonly string[],
-): string | null {
-  const resolvedCandidate = path.resolve(normalizedPath);
-  if (!fs.existsSync(resolvedCandidate)) {
-    return null;
-  }
-
-  let canonicalCandidate: string;
-  try {
-    canonicalCandidate = fs.realpathSync(resolvedCandidate);
-  } catch {
-    return null;
-  }
-
-  for (const root of allowedRoots) {
-    const resolvedRoot = path.resolve(root);
-    const canonicalRoot = fs.existsSync(resolvedRoot)
-      ? fs.realpathSync(resolvedRoot)
-      : resolvedRoot;
-    if (isPathWithinRoot(canonicalCandidate, canonicalRoot)) {
-      return canonicalCandidate;
-    }
-  }
-
-  return null;
-}
-
-function resolveOutboundMediaPath(
-  rawPath: string,
-  mediaKind: QQBotMediaKind,
-  options: ResolveOutboundMediaPathOptions = {},
-): ResolvedOutboundMediaPath {
-  const normalizedPath = normalizePath(rawPath);
-  if (isHttpOrDataSource(normalizedPath)) {
-    return { ok: true, mediaPath: normalizedPath };
-  }
-
-  const allowedPath = resolveQQBotPayloadLocalFilePath(normalizedPath);
-  if (allowedPath) {
-    return { ok: true, mediaPath: allowedPath };
-  }
-
-  if (options.extraLocalRoots && options.extraLocalRoots.length > 0) {
-    const extraAllowedPath = resolveExistingPathWithinRoots(
-      normalizedPath,
-      options.extraLocalRoots,
-    );
-    if (extraAllowedPath) {
-      return { ok: true, mediaPath: extraAllowedPath };
-    }
-  }
-
-  if (options.allowMissingLocalPath) {
-    const allowedMissingPath = resolveMissingPathWithinMediaRoot(normalizedPath);
-    if (allowedMissingPath) {
-      return { ok: true, mediaPath: allowedMissingPath };
-    }
-  }
-
-  debugWarn(`blocked local ${mediaKind} path outside QQ Bot media storage`);
-  return {
-    ok: false,
-    error: `${qqBotMediaKindLabel[mediaKind]} path must be inside QQ Bot media storage`,
-  };
-}
-
-/**
- * Send a photo from a local file, public URL, or Base64 data URL.
- */
-export async function sendPhoto(
-  ctx: MediaTargetContext,
-  imagePath: string,
-): Promise<OutboundResult> {
-  const resolvedMediaPath = resolveOutboundMediaPath(imagePath, "image");
-  if (!resolvedMediaPath.ok) {
-    return { channel: "qqbot", error: resolvedMediaPath.error };
-  }
-  const mediaPath = resolvedMediaPath.mediaPath;
-  const isLocal = isLocalFilePath(mediaPath);
-  const isHttp = mediaPath.startsWith("http://") || mediaPath.startsWith("https://");
-  const isData = mediaPath.startsWith("data:");
-
-  // Force a local download before upload when direct URL upload is disabled.
-  if (isHttp && !shouldDirectUploadUrl(ctx.account)) {
-    debugLog(`sendPhoto: urlDirectUpload=false, downloading URL first...`);
-    const localFile = await downloadToFallbackDir(mediaPath, "sendPhoto");
-    if (localFile) {
-      return await sendPhotoFromLocal(ctx, localFile);
-    }
-    return { channel: "qqbot", error: `Failed to download image: ${mediaPath.slice(0, 80)}` };
-  }
-
-  if (isLocal) {
-    return await sendPhotoFromLocal(ctx, mediaPath);
-  }
-
-  if (!isHttp && !isData) {
-    return { channel: "qqbot", error: `Unsupported image source: ${mediaPath.slice(0, 50)}` };
-  }
-
-  // Remote URL or data: URL — try direct upload first, fall back to
-  // download-then-local on failure.
-  try {
-    const creds = accountToCreds(ctx.account);
-    const target: DeliveryTarget = { type: ctx.targetType, id: ctx.targetId };
-
-    if (target.type === "c2c" || target.type === "group") {
-      const r = await senderSendMedia({
-        target,
-        creds,
-        kind: "image",
-        source: { url: mediaPath },
-        msgId: ctx.replyToId,
-      });
-      return { channel: "qqbot", messageId: r.id, timestamp: r.timestamp };
-    }
-
-
-    if (isHttp) {
-      const r = await senderSendText(target, `![](${mediaPath})`, creds, {
-        msgId: ctx.replyToId,
-      });
-      return { channel: "qqbot", messageId: r.id, timestamp: r.timestamp };
-    }
-    debugLog(`sendPhoto: channel does not support local/Base64 images`);
-    return { channel: "qqbot", error: "Channel does not support local/Base64 images" };
-  } catch (err) {
-    const msg = formatErrorMessage(err);
-
-    // Fall back to plugin-managed download + local upload when QQ fails to
-    // fetch the URL directly. One-shot, non-recursive.
-    if (isHttp && !isData) {
-      debugWarn(
-        `sendPhoto: URL direct upload failed (${msg}), downloading locally and retrying as Base64...`,
-      );
-      const localFile = await downloadToFallbackDir(mediaPath, "sendPhoto");
-      if (localFile) {
-        return await sendPhotoFromLocal(ctx, localFile);
-      }
-    }
-
-    debugError(`sendPhoto failed: ${msg}`);
-    return { channel: "qqbot", error: msg };
-  }
-}
-
-/** Send a photo from a validated local file path. */
-async function sendPhotoFromLocal(
-  ctx: MediaTargetContext,
-  mediaPath: string,
-): Promise<OutboundResult> {
-  if (!(await fileExistsAsync(mediaPath))) {
-    return { channel: "qqbot", error: "Image not found" };
-  }
-  const sizeCheck = checkFileSize(mediaPath, getMaxUploadSize(MediaFileType.IMAGE));
-  if (!sizeCheck.ok) {
-    return buildFileTooLargeResult(MediaFileType.IMAGE, sizeCheck.size);
-  }
-  const mimeType = getImageMimeType(mediaPath);
-  if (!mimeType) {
-    const ext = normalizeLowercaseStringOrEmpty(path.extname(mediaPath));
-    return { channel: "qqbot", error: `Unsupported image format: ${ext}` };
-  }
-  debugLog(`sendPhoto: local (${formatFileSize(sizeCheck.size)})`);
-
-  try {
-    const creds = accountToCreds(ctx.account);
-    const target: DeliveryTarget = { type: ctx.targetType, id: ctx.targetId };
-
-    if (target.type === "c2c" || target.type === "group") {
-      const r = await senderSendMedia({
-        target,
-        creds,
-        kind: "image",
-        source: { localPath: mediaPath },
-        msgId: ctx.replyToId,
-        localPathForMeta: mediaPath,
-      });
-      return { channel: "qqbot", messageId: r.id, timestamp: r.timestamp };
-    }
-    debugLog(`sendPhoto: channel does not support local images`);
-    return { channel: "qqbot", error: "Channel does not support local/Base64 images" };
-  } catch (err) {
-    if (err instanceof UploadDailyLimitExceededError) {
-      debugError(`sendPhoto (local): daily upload quota exceeded`);
-      return buildDailyLimitExceededResult(err);
-    }
-    const msg = formatErrorMessage(err);
-    debugError(`sendPhoto (local) failed: ${msg}`);
-    return { channel: "qqbot", error: msg };
-  }
-}
-
-/**
- * Send voice from either a local file or a public URL.
- *
- * URL handling respects `urlDirectUpload`, and local files are transcoded when needed.
- */
-export async function sendVoice(
-  ctx: MediaTargetContext,
-  voicePath: string,
-  directUploadFormats?: string[],
-  transcodeEnabled: boolean = true,
-): Promise<OutboundResult> {
-  const resolvedMediaPath = resolveOutboundMediaPath(voicePath, "voice", {
-    allowMissingLocalPath: true,
-  });
-  if (!resolvedMediaPath.ok) {
-    return { channel: "qqbot", error: resolvedMediaPath.error };
-  }
-  const mediaPath = resolvedMediaPath.mediaPath;
-  const isHttp = mediaPath.startsWith("http://") || mediaPath.startsWith("https://");
-
-  if (isHttp) {
-    if (shouldDirectUploadUrl(ctx.account)) {
-      try {
-        const creds = accountToCreds(ctx.account);
-        const target: DeliveryTarget = { type: ctx.targetType, id: ctx.targetId };
-        if (target.type === "c2c" || target.type === "group") {
-          const r = await senderSendMedia({
-            target,
-            creds,
-            kind: "voice",
-            source: { url: mediaPath },
-            msgId: ctx.replyToId,
-          });
-          return { channel: "qqbot", messageId: r.id, timestamp: r.timestamp };
-        }
-        debugLog(`sendVoice: voice not supported in channel`);
-        return { channel: "qqbot", error: "Voice not supported in channel" };
-      } catch (err) {
-        const msg = formatErrorMessage(err);
-        debugWarn(
-          `sendVoice: URL direct upload failed (${msg}), downloading locally and retrying...`,
-        );
-      }
-    } else {
-      debugLog(`sendVoice: urlDirectUpload=false, downloading URL first...`);
-    }
-
-    const localFile = await downloadToFallbackDir(mediaPath, "sendVoice");
-    if (localFile) {
-      return await sendVoiceFromLocal(ctx, localFile, directUploadFormats, transcodeEnabled);
-    }
-    return { channel: "qqbot", error: `Failed to download audio: ${mediaPath.slice(0, 80)}` };
-  }
-
-  return await sendVoiceFromLocal(ctx, mediaPath, directUploadFormats, transcodeEnabled);
-}
-
-/** Send voice from a local file. */
-async function sendVoiceFromLocal(
-  ctx: MediaTargetContext,
-  mediaPath: string,
-  directUploadFormats: string[] | undefined,
-  transcodeEnabled: boolean,
-): Promise<OutboundResult> {
-  // TTS can still be flushing the file to disk, so wait for a stable file first.
-  const fileSize = await waitForFile(mediaPath);
-  if (fileSize === 0) {
-    return { channel: "qqbot", error: "Voice generate failed" };
-  }
-  if (fileSize > getMaxUploadSize(MediaFileType.VOICE)) {
-    return buildFileTooLargeResult(MediaFileType.VOICE, fileSize);
-  }
-
-  // Re-check containment after the file appears to prevent symlink-race escapes.
-  const safeMediaPath = resolveQQBotPayloadLocalFilePath(mediaPath);
-  if (!safeMediaPath) {
-    debugWarn(`sendVoice: blocked local voice path outside QQ Bot media storage`);
-    return { channel: "qqbot", error: "Voice path must be inside QQ Bot media storage" };
-  }
-
-  const needsTranscode = shouldTranscodeVoice(safeMediaPath);
-
-  if (needsTranscode && !transcodeEnabled) {
-    const ext = normalizeLowercaseStringOrEmpty(path.extname(safeMediaPath));
-    debugLog(
-      `sendVoice: transcode disabled, format ${ext} needs transcode, returning error for fallback`,
-    );
-    return {
-      channel: "qqbot",
-      error: `Voice transcoding is disabled and format ${ext} cannot be uploaded directly`,
-    };
-  }
-
-  try {
-    const silkBase64 = await audioFileToSilkBase64(safeMediaPath, directUploadFormats);
-    let uploadBase64 = silkBase64;
-
-    if (!uploadBase64) {
-      const buf = await readFileAsync(safeMediaPath);
-      uploadBase64 = buf.toString("base64");
-      debugLog(`sendVoice: SILK conversion failed, uploading raw (${formatFileSize(buf.length)})`);
-    } else {
-      debugLog(`sendVoice: SILK ready (${fileSize} bytes)`);
-    }
-
-    const creds = accountToCreds(ctx.account);
-    const target: DeliveryTarget = { type: ctx.targetType, id: ctx.targetId };
-
-    if (target.type === "c2c" || target.type === "group") {
-      const r = await senderSendMedia({
-        target,
-        creds,
-        kind: "voice",
-        source: { base64: uploadBase64 },
-        msgId: ctx.replyToId,
-        localPathForMeta: safeMediaPath,
-      });
-      return { channel: "qqbot", messageId: r.id, timestamp: r.timestamp };
-    }
-    debugLog(`sendVoice: voice not supported in channel`);
-    return { channel: "qqbot", error: "Voice not supported in channel" };
-  } catch (err) {
-    if (err instanceof UploadDailyLimitExceededError) {
-      debugError(`sendVoice (local): daily upload quota exceeded`);
-      return buildDailyLimitExceededResult(err);
-    }
-    const msg = formatErrorMessage(err);
-    debugError(`sendVoice (local) failed: ${msg}`);
-    return { channel: "qqbot", error: msg };
-  }
-}
-
-/** Send video from either a public URL or a local file. */
-export async function sendVideoMsg(
-  ctx: MediaTargetContext,
-  videoPath: string,
-): Promise<OutboundResult> {
-  const resolvedMediaPath = resolveOutboundMediaPath(videoPath, "video");
-  if (!resolvedMediaPath.ok) {
-    return { channel: "qqbot", error: resolvedMediaPath.error };
-  }
-  const mediaPath = resolvedMediaPath.mediaPath;
-  const isHttp = mediaPath.startsWith("http://") || mediaPath.startsWith("https://");
-
-  if (isHttp && !shouldDirectUploadUrl(ctx.account)) {
-    debugLog(`sendVideoMsg: urlDirectUpload=false, downloading URL first...`);
-    const localFile = await downloadToFallbackDir(mediaPath, "sendVideoMsg");
-    if (localFile) {
-      return await sendVideoFromLocal(ctx, localFile);
-    }
-    return { channel: "qqbot", error: `Failed to download video: ${mediaPath.slice(0, 80)}` };
-  }
-
-  try {
-    if (isHttp) {
-      const creds = accountToCreds(ctx.account);
-      const target: DeliveryTarget = { type: ctx.targetType, id: ctx.targetId };
-      if (target.type === "c2c" || target.type === "group") {
-        const r = await senderSendMedia({
-          target,
-          creds,
-          kind: "video",
-          source: { url: mediaPath },
-          msgId: ctx.replyToId,
-        });
-        return { channel: "qqbot", messageId: r.id, timestamp: r.timestamp };
-      }
-      debugLog(`sendVideoMsg: video not supported in channel`);
-      return { channel: "qqbot", error: "Video not supported in channel" };
-    }
-
-    return await sendVideoFromLocal(ctx, mediaPath);
-  } catch (err) {
-    const msg = formatErrorMessage(err);
-
-    // If direct URL upload fails, retry through a local download path.
-    if (isHttp) {
-      debugWarn(
-        `sendVideoMsg: URL direct upload failed (${msg}), downloading locally and retrying as Base64...`,
-      );
-      const localFile = await downloadToFallbackDir(mediaPath, "sendVideoMsg");
-      if (localFile) {
-        return await sendVideoFromLocal(ctx, localFile);
-      }
-    }
-
-    debugError(`sendVideoMsg failed: ${msg}`);
-    return { channel: "qqbot", error: msg };
-  }
-}
-
-/** Send video from a local file. */
-async function sendVideoFromLocal(
-  ctx: MediaTargetContext,
-  mediaPath: string,
-): Promise<OutboundResult> {
-  if (!(await fileExistsAsync(mediaPath))) {
-    return { channel: "qqbot", error: "Video not found" };
-  }
-  const sizeCheck = checkFileSize(mediaPath, getMaxUploadSize(MediaFileType.VIDEO));
-  if (!sizeCheck.ok) {
-    return buildFileTooLargeResult(MediaFileType.VIDEO, sizeCheck.size);
-  }
-  debugLog(`sendVideoMsg: local video (${formatFileSize(sizeCheck.size)})`);
-
-  try {
-    const creds = accountToCreds(ctx.account);
-    const target: DeliveryTarget = { type: ctx.targetType, id: ctx.targetId };
-    if (target.type === "c2c" || target.type === "group") {
-      const r = await senderSendMedia({
-        target,
-        creds,
-        kind: "video",
-        source: { localPath: mediaPath },
-        msgId: ctx.replyToId,
-        localPathForMeta: mediaPath,
-      });
-      return { channel: "qqbot", messageId: r.id, timestamp: r.timestamp };
-    }
-    debugLog(`sendVideoMsg: video not supported in channel`);
-    return { channel: "qqbot", error: "Video not supported in channel" };
-  } catch (err) {
-    if (err instanceof UploadDailyLimitExceededError) {
-      debugError(`sendVideoMsg (local): daily upload quota exceeded`);
-      return buildDailyLimitExceededResult(err);
-    }
-    const msg = formatErrorMessage(err);
-    debugError(`sendVideoMsg (local) failed: ${msg}`);
-    return { channel: "qqbot", error: msg };
-  }
-}
-
-/** Send a file from a local path or public URL. */
-export async function sendDocument(
-  ctx: MediaTargetContext,
-  filePath: string,
-  options: SendDocumentOptions = {},
-): Promise<OutboundResult> {
-  const extraLocalRoots = options.allowQQBotDataDownloads
-    ? [getQQBotDataDir("downloads")]
-    : undefined;
-  const resolvedMediaPath = resolveOutboundMediaPath(filePath, "file", {
-    extraLocalRoots,
-  });
-  if (!resolvedMediaPath.ok) {
-    return { channel: "qqbot", error: resolvedMediaPath.error };
-  }
-  const mediaPath = resolvedMediaPath.mediaPath;
-  const isHttp = mediaPath.startsWith("http://") || mediaPath.startsWith("https://");
-  const fileName = sanitizeFileName(path.basename(mediaPath));
-
-  if (isHttp && !shouldDirectUploadUrl(ctx.account)) {
-    debugLog(`sendDocument: urlDirectUpload=false, downloading URL first...`);
-    const localFile = await downloadToFallbackDir(mediaPath, "sendDocument");
-    if (localFile) {
-      return await sendDocumentFromLocal(ctx, localFile);
-    }
-    return { channel: "qqbot", error: `Failed to download file: ${mediaPath.slice(0, 80)}` };
-  }
-
-  try {
-    if (isHttp) {
-      const creds = accountToCreds(ctx.account);
-      const target: DeliveryTarget = { type: ctx.targetType, id: ctx.targetId };
-      if (target.type === "c2c" || target.type === "group") {
-        const r = await senderSendMedia({
-          target,
-          creds,
-          kind: "file",
-          source: { url: mediaPath },
-          msgId: ctx.replyToId,
-          fileName,
-        });
-        return { channel: "qqbot", messageId: r.id, timestamp: r.timestamp };
-      }
-      debugLog(`sendDocument: file not supported in channel`);
-      return { channel: "qqbot", error: "File not supported in channel" };
-    }
-
-    return await sendDocumentFromLocal(ctx, mediaPath);
-  } catch (err) {
-    const msg = formatErrorMessage(err);
-
-    // If direct URL upload fails, retry through a local download path.
-    if (isHttp) {
-      debugWarn(
-        `sendDocument: URL direct upload failed (${msg}), downloading locally and retrying as Base64...`,
-      );
-      const localFile = await downloadToFallbackDir(mediaPath, "sendDocument");
-      if (localFile) {
-        return await sendDocumentFromLocal(ctx, localFile);
-      }
-    }
-
-    debugError(`sendDocument failed: ${msg}`);
-    return { channel: "qqbot", error: msg };
-  }
-}
-
-/** Send a file from local storage. */
-async function sendDocumentFromLocal(
-  ctx: MediaTargetContext,
-  mediaPath: string,
-): Promise<OutboundResult> {
-  const fileName = sanitizeFileName(path.basename(mediaPath));
-
-  if (!(await fileExistsAsync(mediaPath))) {
-    return { channel: "qqbot", error: "File not found" };
-  }
-  const sizeCheck = checkFileSize(mediaPath, getMaxUploadSize(MediaFileType.FILE));
-  if (!sizeCheck.ok) {
-    return buildFileTooLargeResult(MediaFileType.FILE, sizeCheck.size);
-  }
-  if (sizeCheck.size === 0) {
-    return { channel: "qqbot", error: `File is empty: ${mediaPath}` };
-  }
-  debugLog(`sendDocument: local file (${formatFileSize(sizeCheck.size)})`);
-
-  try {
-    const creds = accountToCreds(ctx.account);
-    const target: DeliveryTarget = { type: ctx.targetType, id: ctx.targetId };
-    if (target.type === "c2c" || target.type === "group") {
-      const r = await senderSendMedia({
-        target,
-        creds,
-        kind: "file",
-        source: { localPath: mediaPath },
-        msgId: ctx.replyToId,
-        fileName,
-        localPathForMeta: mediaPath,
-      });
-      return { channel: "qqbot", messageId: r.id, timestamp: r.timestamp };
-    }
-    debugLog(`sendDocument: file not supported in channel`);
-    return { channel: "qqbot", error: "File not supported in channel" };
-  } catch (err) {
-    if (err instanceof UploadDailyLimitExceededError) {
-      debugError(`sendDocument (local): daily upload quota exceeded`);
-      return buildDailyLimitExceededResult(err);
-    }
-    const msg = formatErrorMessage(err);
-    debugError(`sendDocument (local) failed: ${msg}`);
-    return { channel: "qqbot", error: msg };
-  }
-}
-
-/** Download a remote file into the fallback media directory. */
-async function downloadToFallbackDir(httpUrl: string, caller: string): Promise<string | null> {
-  try {
-    const downloadDir = getQQBotMediaDir("downloads", "url-fallback");
-    const localFile = await downloadFile(httpUrl, downloadDir);
-    if (!localFile) {
-      debugError(`${caller} fallback: download also failed for ${httpUrl.slice(0, 80)}`);
-      return null;
-    }
-    debugLog(`${caller} fallback: downloaded → ${localFile}`);
-    return localFile;
-  } catch (err) {
-    debugError(`${caller} fallback download error:`, err);
-    return null;
-  }
-}
+const isImageFile = coreIsImageFile;
+const isVideoFile = coreIsVideoFile;
 
 /**
  * Send text, optionally falling back from passive reply mode to proactive mode.
@@ -953,7 +134,6 @@ export async function sendText(ctx: OutboundContext): Promise<OutboundResult> {
   if (mediaTagMatches && mediaTagMatches.length > 0) {
     debugLog(`[qqbot] sendText: Detected ${mediaTagMatches.length} media tag(s), processing...`);
 
-    // Preserve the original text/media ordering when sending mixed content.
     const sendQueue: Array<{
       type: "text" | "image" | "voice" | "video" | "file" | "media";
       content: string;
@@ -981,12 +161,8 @@ export async function sendText(ctx: OutboundContext): Promise<OutboundResult> {
       }
       mediaPath = normalizePath(mediaPath);
 
-      // Fix paths that the model emitted with markdown-style escaping.
       mediaPath = mediaPath.replace(/\\\\/g, "\\");
 
-      // Skip octal escape decoding for Windows local paths (e.g. C:\Users\1\file.txt)
-      // where backslash-digit sequences like \1, \2 ... \7 are directory separators,
-      // not octal escape sequences.
       const isWinLocal = /^[a-zA-Z]:[\\/]/.test(mediaPath) || mediaPath.startsWith("\\\\");
       try {
         const hasOctal = /\\[0-7]{1,3}/.test(mediaPath);
@@ -996,7 +172,7 @@ export async function sendText(ctx: OutboundContext): Promise<OutboundResult> {
           debugLog(`[qqbot] sendText: Decoding path with mixed encoding: ${mediaPath}`);
 
           let decoded = mediaPath.replace(/\\([0-7]{1,3})/g, (_: string, octal: string) => {
-            return String.fromCharCode(Number.parseInt(octal, 8));
+            return String.fromCharCode(parseInt(octal, 8));
           });
 
           const bytes: number[] = [];
@@ -1058,7 +234,6 @@ export async function sendText(ctx: OutboundContext): Promise<OutboundResult> {
 
     debugLog(`[qqbot] sendText: Send queue: ${sendQueue.map((item) => item.type).join(" -> ")}`);
 
-    // Send queue items in order.
     const mediaTarget = buildMediaTarget({ to, account, replyToId });
     let lastResult: OutboundResult = { channel: "qqbot" };
 
@@ -1098,7 +273,6 @@ export async function sendText(ctx: OutboundContext): Promise<OutboundResult> {
         } else if (item.type === "file") {
           lastResult = await sendDocument(mediaTarget, item.content);
         } else if (item.type === "media") {
-          // Auto-route qqmedia based on the file extension.
           lastResult = await sendMedia({
             to,
             text: "",
@@ -1178,6 +352,7 @@ export async function sendMedia(ctx: MediaOutboundContext): Promise<OutboundResu
   if (!ctx.mediaUrl) {
     return { channel: "qqbot", error: "mediaUrl is required for sendMedia" };
   }
+
   const resolvedMediaPath = resolveOutboundMediaPath(ctx.mediaUrl, "media", {
     allowMissingLocalPath: true,
   });
@@ -1188,8 +363,6 @@ export async function sendMedia(ctx: MediaOutboundContext): Promise<OutboundResu
 
   const target = buildMediaTarget({ to, account, replyToId });
 
-  // Dispatch by type, preferring MIME and falling back to the file extension.
-  // Individual send* helpers already handle direct URL upload vs. download fallback.
   if (isAudioFile(mediaUrl, mimeType)) {
     const formats =
       account.config?.audioFormatPolicy?.uploadDirectFormats ??
@@ -1202,7 +375,6 @@ export async function sendMedia(ctx: MediaOutboundContext): Promise<OutboundResu
       }
       return result;
     }
-    // Preserve the voice error and fall back to file send.
     const voiceError = result.error;
     debugWarn(`[qqbot] sendMedia: sendVoice failed (${voiceError}), falling back to sendDocument`);
     const fallback = await sendDocument(target, mediaUrl);
@@ -1223,7 +395,6 @@ export async function sendMedia(ctx: MediaOutboundContext): Promise<OutboundResu
     return result;
   }
 
-  // Non-image, non-audio, and non-video media fall back to file send.
   if (
     !isImageFile(mediaUrl, mimeType) &&
     !isAudioFile(mediaUrl, mimeType) &&
@@ -1236,7 +407,6 @@ export async function sendMedia(ctx: MediaOutboundContext): Promise<OutboundResu
     return result;
   }
 
-  // Default to image handling. sendPhoto already contains URL fallback logic.
   const result = await sendPhoto(target, mediaUrl);
   if (!result.error && text?.trim()) {
     await sendTextAfterMedia(target, text);
@@ -1244,7 +414,6 @@ export async function sendMedia(ctx: MediaOutboundContext): Promise<OutboundResu
   return result;
 }
 
-/** Send text after media when the transport supports a follow-up text message. */
 async function sendTextAfterMedia(ctx: MediaTargetContext, text: string): Promise<void> {
   try {
     const creds = accountToCreds(ctx.account);
@@ -1255,22 +424,6 @@ async function sendTextAfterMedia(ctx: MediaTargetContext, text: string): Promis
   }
 }
 
-// Media type detection delegated to core/outbound/media-type-detect.ts.
-// Re-alias for backward compatibility within this file.
-const isImageFile = coreIsImageFile;
-const isVideoFile = coreIsVideoFile;
-
-/**
- * Send a proactive (no reply context) text message to a qualified target.
- *
- * Thin wrapper around {@link sendText} for callers that have a fully-qualified
- * target string (e.g. `"qqbot:c2c:<openid>"`) and a {@link GatewayAccount},
- * and do not want to manage access tokens or delivery-target parsing manually.
- *
- * @param account Resolved gateway account.
- * @param to Fully-qualified target address (`qqbot:c2c:<openid>`, `qqbot:group:<id>`, etc.).
- * @param content Message content.
- */
 export async function sendProactiveMessage(
   account: GatewayAccount,
   to: string,
@@ -1279,31 +432,6 @@ export async function sendProactiveMessage(
   return sendText({ account, to, text: content });
 }
 
-/**
- * Send a message emitted by an OpenClaw cron task.
- *
- * Cron output may be either:
- * 1. A `QQBOT_CRON:{base64}` structured payload that includes target metadata.
- * 2. Plain text that should be sent directly to the provided fallback target.
- *
- * @param account Resolved account configuration.
- * @param to Fallback target address when the payload does not include one.
- * @param message Message content, either `QQBOT_CRON:` payload or plain text.
- * @returns Send result.
- *
- * @example
- * ```typescript
- * // Structured payload
- * const result = await sendCronMessage(
- *   account,
- *   "user_openid",
- *   "QQBOT_CRON:eyJ0eXBlIjoiY3Jvbl9yZW1pbmRlciIs..."
- * );
- *
- * // Plain text
- * const result = await sendCronMessage(account, "user_openid", "This is a plain reminder message.");
- * ```
- */
 export async function sendCronMessage(
   account: GatewayAccount,
   to: string,
@@ -1312,7 +440,6 @@ export async function sendCronMessage(
   const timestamp = new Date().toISOString();
   debugLog(`[${timestamp}] [qqbot] sendCronMessage: to=${to}, message length=${message.length}`);
 
-  // Detect `QQBOT_CRON:` structured payloads first.
   const cronResult = decodeCronPayload(message);
 
   if (cronResult.isCronPayload) {
@@ -1332,7 +459,6 @@ export async function sendCronMessage(
         `[${timestamp}] [qqbot] sendCronMessage: decoded cron payload, targetType=${payload.targetType}, targetAddress=${payload.targetAddress}, content length=${payload.content.length}`,
       );
 
-      // Prefer the target encoded in the structured payload.
       const targetTo =
         payload.targetType === "group" ? `group:${payload.targetAddress}` : payload.targetAddress;
 
@@ -1340,7 +466,6 @@ export async function sendCronMessage(
         `[${timestamp}] [qqbot] sendCronMessage: sending proactive message to targetTo=${targetTo}`,
       );
 
-      // Send the reminder content.
       const result = await sendText({ account, to: targetTo, text: payload.content });
 
       if (result.error) {
@@ -1355,7 +480,6 @@ export async function sendCronMessage(
     }
   }
 
-  // Fall back to plain text handling when the payload is not structured.
   debugLog(`[${timestamp}] [qqbot] sendCronMessage: plain text message, sending to ${to}`);
   return await sendText({ account, to, text: message });
 }
